requirements.txt
# --- Evaluation ---
sacrebleu>=2.4.0      # Official chrF++ metric

# --- Inference & Embeddings ---
# Use LiteLLM to talk to Ollama (Inference) AND Ollama (Embeddings)
litellm>=1.35.0

# --- Vector Store (RAG) ---
faiss-cpu>=1.8.0

# --- Data & Logic ---
pandas>=2.0.0         # For handling .jsonl and training data
numpy>=1.24.0         # Needed for vector math with FAISS
python-dotenv>=1.0.0  # Manages local paths/API keys without hardcoding
pyyaml>=6.0.0         # For your experiment_config.yaml
tqdm>=4.66.0          # Progress bars for the 90 test instances